feat-#67: move getAuthenticatedUser to user module
