+++ src/auth/authorization/services/authorization.service.ts
@@ -5,9 +5,6 @@ import { AccessControlDto } from '../dto';
 import { IJwtPayload } from '../../passport-jwt/interfaces';
 import { RpcExceptionService } from '../../../utils/exception-handling';
 import { SetRefreshTokenCommand } from '../commands/impl';
-import { AuthCredentialsDto } from '../../authentication/dto';
-import { UserService } from '../../user/services/user.service';
-import { PasswordUtilsService } from '../../../utils/password-utils';
 import { CookieService } from '../../authentication/services';
 import { TokenService } from './token.service';
 
@@ -16,8 +13,6 @@ export class AuthorizationService {
   constructor(
     private readonly commandBus: CommandBus,
     private readonly rpcExceptionService: RpcExceptionService,
-    private readonly userService: UserService,
-    private readonly passwordUtilsService: PasswordUtilsService,
     private readonly cookieService: CookieService,
     private readonly tokenService: TokenService,
   ) {}
@@ -45,23 +40,5 @@ export class AuthorizationService {
 
   getCookieWithJwtAccessToken(id: number): string {
     return this.cookieService.getCookieWithJwtAccessToken(id);
-  }
-
-  async getAuthenticatedUser(authCredentialsDto: AuthCredentialsDto) {
-    const user = await this.userService.getUserByEmail(
-      authCredentialsDto.email,
-    );
-
-    if (!user) this.rpcExceptionService.throwNotFound('User not found');
-
-    const isPasswordValidated = await this.passwordUtilsService.validatePassword(
-      authCredentialsDto.password,
-      user.password,
-    );
-
-    if (!isPasswordValidated)
-      this.rpcExceptionService.throwUnauthorised('Password do not match');
-
-    return user;
   }
 }
